Improve migration to delete ACL, Relationships

+++ src/ggrc/migrations/versions/20180706140827_0e40a37a05f4_add_foreignkey_snapshot.py
@@ -18,23 +18,59 @@ revision = '0e40a37a05f4'
 down_revision = '054d15be7a29'
 
 
+COLLECT_BAD_SNAPSHOT_IDS = """
+    INSERT INTO `tmp_bad_snapshots`
+    SELECT `s`.`id` AS `id`
+    FROM `snapshots` AS `s`
+    LEFT JOIN `audits` AS `a`
+    ON `s`.`parent_id` = `a`.`id`
+    WHERE `a`.`id` IS null OR `s`.`parent_type` != "Audit"
+"""
+
+
 def upgrade():
   """Upgrade database schema and/or data, creating a new revision."""
-  op.execute(
-      """
-      DELETE FROM snapshots
-      WHERE id IN (
-          SELECT id
-          FROM (
-              SELECT s.id as id
-              FROM snapshots as s
-              LEFT JOIN audits as a
-              ON s.parent_id = a.id
-              WHERE a.id IS null OR s.parent_type != "Audit"
-          ) sq
-      );
-      """
+  connection = op.get_bind()
+  connection.execute(
+      "CREATE TEMPORARY TABLE `tmp_bad_snapshots`(`id` INT(11))"
   )
+  if connection.execute(COLLECT_BAD_SNAPSHOT_IDS).rowcount:
+    connection.execute(
+        """
+        DELETE FROM snapshots
+        WHERE id IN (SELECT `id` FROM `tmp_bad_snapshots`)
+        """
+    )
+    connection.execute(
+        """
+        DELETE FROM access_control_list
+        WHERE
+            object_id IN (SELECT `id` FROM `tmp_bad_snapshots`)
+            AND
+            object_type = "Snapshot"
+        """
+    )
+    connection.execute(
+        """
+        DELETE FROM relationships
+        WHERE
+            source_id IN (SELECT `id` FROM `tmp_bad_snapshots`)
+            AND
+            source_type = "Snapshot";
+        """
+    )
+    # Because temporary table can't be used twice in one query, deletion
+    # of relationships made in 2 queries.
+    connection.execute(
+        """
+        DELETE FROM relationships
+        WHERE
+            destination_id IN (SELECT `id` FROM `tmp_bad_snapshots`)
+            AND
+            destination_type = "Snapshot";
+        """
+    )
+  connection.execute("DROP TABLE IF EXISTS `tmp_bad_snapshots`")
   op.create_foreign_key(
       'fk_snapshots_audits',
       'snapshots',
